Add default values
Section 1a is only for classroom purposes. lines 14 - 16 SHOULD never
be used in production. Line 12 set to false is ok for production

Section 1c is where you declare a variable for every form element you
have. It is here you set the form variable to its default value. While
you are testing it is just easier to have the email address set to
yours so you don’t have to type it. Be sure when you go live to set it
to nothing, just quote quote (“”)

Section 3d line 305 shows how we print the default value on the html
form. You would do this (or something like this) for every form object
with the variables declared in section 1c.